initrd: set a default carrier timeout of 15 seconds in initrd
In initrd, a too low carrier timeout means that the machine will
possibly fail to boot. On the other hand, increasing the value doesn't
have side effects, except for a bit longer delay on some machines.

Increase the value to 15 seconds. Note that the default value is not
propagated to the real root.