ppp: prevent commits on interface while IPCP/IPV6CP is running
pppd also tries to configure addresses by itself, and quits if it
fails to do so. If we commit addresses on the device at the same time,
we might race with pppd. So, we are not allowed to touch the device
until pppd sends us an IP configuration, which happens after it has
finished configuring the interface.

https://bugzilla.redhat.com/show_bug.cgi?id=2085382